am64x/am243x: sdl: Fix VTM temperature datatype
- VTM is capable of reading negative celsius

- Range is roughly -49 to 160 C

- VTM driver was typecasting to uint32_t

- Issues in negative temperature configuration

- Fix by changing to int32_t

- V1 driver and V1 devices are outside scope of this fix

Fixes: PROC_SDL-9927

+++ source/sdl/vtm/v0/sdl_vtm.c
@@ -196,7 +196,7 @@ int32_t SDL_VTM_initTs(SDL_VTM_InstTs instance, const SDL_VTM_configTs *pConfig)
  /**
  * Design: PROC_SDL-1165,PROC_SDL-1303,PROC_SDL-1304
  */
-int32_t SDL_VTM_getTemp(SDL_VTM_InstTs instance, uint32_t *pTempVal)
+int32_t SDL_VTM_getTemp(SDL_VTM_InstTs instance, int32_t *pTempVal)
 {
     const SDL_VTM_cfg1Regs      *p_cfg1;
 	uint32_t baseAddr;
@@ -218,7 +218,7 @@ int32_t SDL_VTM_getTemp(SDL_VTM_InstTs instance, uint32_t *pTempVal)
     {
         adc_code = SDL_VTM_getAdcCode(p_sensor);
         sdlResult = SDL_VTM_tsConvADCToTemp (adc_code, instance, &p_milli_degree_temp_val);
-        *pTempVal = (uint32_t) p_milli_degree_temp_val;
+        *pTempVal = p_milli_degree_temp_val;
     }
     return (sdlResult);
 }
